Array
Q1. Consider Array of 5 elements, print address of 5 elements. Take address of 0th element and calculate address of other elements manually.

Q2. Calculate sum of array elements using function and without function. Do the dry run as well.

Q3. Execute the program and try dry run for a[2], *a+2, a+2, *(a+2)